Refactor: 식물도감 페이지 컴포넌트에서 API 관련 함수 추출
